main create post screen

+++ lib/domain/models/request/create_post_request.dart
@@ -29,7 +29,7 @@ class CreatePostRequest {
   int? hasAudio;
   String? imageUrl;
   String? fileName;
-  double? size;
+  int? size;
   int? scheduleTime;
   String? cloudinaryPublicId;
